requirements.txt
attrs==17.2.0
Pillow==4.3.0
